backend/app/kafka_producer.py
import json
import os
from kafka import KafkaProducer

KAFKA_BROKER = os.getenv("KAFKA_BROKER", "kafka:9092")
TOPIC = os.getenv("KAFKA_VITALS_TOPIC", "vitals-stream")

producer = KafkaProducer(
    bootstrap_servers=KAFKA_BROKER,
    value_serializer=lambda v: json.dumps(v).encode("utf-8"),
)


def publish_vital(message: dict):
    producer.send(TOPIC, message)
    producer.flush()
